galp_airflow/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
apache-airflow-providers-cncf-kubernetes
dbt-core
dbt-postgres
psycopg2-binary


